wind

+++ src/globals.c
@@ -1,7 +1,7 @@
 #include "globals.h"
 
-f32 TIME = 0;
-f32 FRAME_TIME = 0;
-f32 DELTA_TIME = 0;
+f64 TIME = 0;
+f64 FRAME_TIME = 0;
+f64 DELTA_TIME = 0;
 u32 PREVIOUS_TIME = 0;
 u32 CURRENT_TIME = 0;
